Remove kwargs on solver registration (#20)
The API is now much closer to CVXPY's. Using `cvxpy-gurobi` is now as
simple as changing:
```diff
- problem.solve(solver=cp.GUROBI, QCPDual=True)
+ problem.solve(method=GUROBI_TRANSLATION, QCPDual=True)
```
or without registration:
```diff
- problem.solve(solver=cp.GUROBI, QCPDual=True)
+ cvxpy_gurobi.solve(problem, QCPDual=True)
```

To keep things simple and make it easier to change in the future, a few
functions were removed from the public API, and the README was updated
to account for these changes.