Create Gurobi variables during translation (#24)
As opposed to having a specific step before translation. It helps to
make variables similar to other nodes. This will be necessary to support
more complex expressions #23 as some expressions' translation may
require additional variables.

Removing the `map_variables` step revealed that separating the objective
and constraint translaters was unnecessary, as now any translation
method may add variables/constraints to the model.

One downside is that the translation can no longer be run "model-free"
to generate modle-agnostic expressions. This impression was wrong
however as any variable used in these expressions was linked to a Gurobi
model anyway.