Support `reshape` atom (#42)
This atom generates an auxilliary variable in many cases where it is not
necessary, so perhaps we could have some way to indicate whether reshape
should be strict or not. An example of a strict reshape is going from ()
to (1,). Both give the same result in most situations, but maybe not
all. Until we have a good coverage of cases where strictness is required
(e.g. when stacking), I prefer to be strict by default.

Contributes to #32.