Stack is not available on `gurobipy<11` (#44)
This PR updates the test matrix to get all tests passing. Noticeable
changes:
- `gp.{h,v}stack` was added in version 11, so skip the corresponding
tests and raise an error during translation,
- older versions' tests are run first as they are the most likely to
have failures,
- removed some version restrictions on numpy since the issues were fixed
in recent versions,
- `pytest-xdist` is now installed in test environments, the number of
tests grew large enough that the time spent instantiating workers is
smaller than the gain due to parallelization.

There is a weird issue with `cvxpy==1.4`, not sure what it's about yet,
so it's not tested.