Ensure `minimum`/`maximum` arguments are scalars (#51)
Fixes #47 by forcing the `gp.min` and `gp.max` arguments to be scalars.
I added a new parameter to `translate_into_variable` to explicitly ask
for a scalar variable. Before this change, we tried to create a scalar
variable when possible, but now it's explicit.

Now the responsabilities of `translate_into_variable` and
`translate_into_scalar` are a little mixed, so it might be worth
revisiting/merging them in the future.

The snapshot differences come from the change in the `_min_max` function
to not create an auxilliary variable if the argument is already scalar.
It makes a nicer looking problem, but we lose track of the `min`/`max`
completely, so I'm still not 100% sure it should stay this way.