Add lower bounds to dependencies and test them (#56)
Add a new environment, "oldest", using UV's lowest resolution strategy.
This allows us to test the oldest versions we can support.

The lower bounds were set to the first versions of `numpy`, `scipy` and
`cvxpy` that could be installed together (at least on my machine) with
Python 3.8. They are intentionally very outdated, but almost no changes
are needed to support them so why not?

One noticeable change is that the versions of pytest and pytest-insta
had to be downgraded to work with Python 3.8, and the name of the
snapshot files changed a little.