Avoid testing on `gurobipy` 10 since the license expired (#68)
It expired on 2024-10-28, so none of the tests are passing with that
version.

Bumping to v11 unearth another issue with `vstack` (which is skipped on
older versions of `gurobipy`), this time related to the version of
`cvxpy` used in the _oldest_ tests. So I also bumped the minimum tested
version of `cvxpy` to 1.2.0 (which released cvxpy/cvxpy#1339), and
switched to `cvxpy-base`.

The testing scope of `gurobipy` is now limited to v11, but v12 should be
out _soon_ anyway so it won't be the newest version anymore.