Set up layout for SCIP (#132)
It's a copy of the code used for Gurobi, with renamings and the bare
minimum to be able to import the package.

I also started the testing infrastructure, so that tests can be enabled
as features are added.

The minimum version of `pyscipopt` was set to 5.4.1 as that is the first
version where the `MatrixVariable` was made available. I suspect that we
will need to upgrade to 5.5.0 because a lot of changes were made to
improve/fix the matrix API.