Support h- and v-stack with SCIP (#154)
Since the SCIP matrix API just uses NumPy arrays, it is really easy.
Though the result of the stack is not a SCIP object, it's a plain array.
It doesn't seem to cause issues when computing, but there might be some
edge cases where it fails?
Oh and this greatly improved the `quad_form_stack` test case as we don't
need the auxiliary variable anymore :tada: