Use common exception hierarchy (#160)
Move the base exceptions to the library root, which makes it possible to
catch exceptions no matter which solver/translater is used.
This enables us to unblock the tests for invalid cases for both SCIP and
Gurobi.

There is something weird going on with gurobipy's `hstack`, not sure
why. There's no issue with `vstack`. But for now we just skip both on
Gurobi.