Add global interface that dispatches to the solvers (#163)
The idea is to be able to use the high-level interface without importing
from a solver subpackage. The high-level interface then simply redirects
to the appropriate function in the solver packages.

Care was taken to make sure all files in the global package can be
imported without having `gurobipy` or `pyscipopt` installed. Then we
will try to import the solvers only once one is requested explicitly.